+++ core/__init__.py
@@ -1,4 +1,4 @@
 from .config import Config, ConfigError
-from .kapellmeister import Kapellmeister
+from .kapellmeister import Kapellmeister, TimeoutException
 
-__all__ = ["Config", "ConfigError", "Kapellmeister"]
+__all__ = ["Config", "ConfigError", "Kapellmeister", "TimeoutException"]